refactor: remote -> service 패키지명 수정
