+++ app/src/main/java/woowacourse/shopping/data/service/CouponService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote
+package woowacourse.shopping.data.service
 
 import retrofit2.Response
 import retrofit2.http.GET
